Fix query box on first visit, add autosave
The query box now shows up on the first visit before a cookie is set.

Program and query and saved automatically every 60 seconds just like
when the user closes the site. This is to ensure a browser crash
(which will not trigger the save event) doesn't make the user lose
their data.