Wasm: Fix memory not freed after OOM
An error message is now reliably thrown when OOM happens.
Also, memory is being freed after to ensure the app keeps working.
Yet there is a memory leak that when refreshing a page, old allocated
memory does not get freed and more is allocated above the limit.
That is likely a bug outside this project.

Reduced memory limit to 500MB. 1GB is a bit tight on a 4GB RAM
machine when running a bit more in the background. 

UI: Changed back to results coming from top for now. Scrollbar did not
appear in the previous version.