Cleanup of wasm/web files; fix build files 404
Goal was to let emrun serve from web directory only, so no other files
are exposed through the webserver and clutter in the root of the
project is minimized. Since the last modifications of the wasm
build system, build output is now in the *build* directory,
so outside the server's document root. Thus we just symlink the build
directory into the document root of the server and be happy.