Fix calling clear() did not free all memory
Calling bfs_organizer.clear() was supposed to free up all memory used
by the bfs_organizer object, e.g. when clicking on the X in the
frontend next to a result.
However, this was not working. The problem is that clearing all
references to a c++ object in JS won't call its destructors
(https://emscripten.org/docs/porting/connecting_cpp_and_javascript/embind.html#memory-management).
The idea was: if JS can't call the destructors, why don't we call
them in the c++ code. To do that, I changed bfs_organizer to be
on the heap, rather than the stack, which enables me to call
delete on it. That in turn, calls all destructors.

I suspect that this leaves a little bit of garbage in memory,
namely the PrologBFSWasmWrapper object. JS still can't call the
destructor of that object, but that shouldn't really be a big
issue.