Allow catch tests in browser, cleanup cmake
This commit aims to allow running catch tests in the browser, which
puts them closer to the real world usage. This was done primarily
because we wanted to make sure that using -O3 optimizations in
emscripten won't break anything.

Along the way I also decided that it would be a good idea to update
our cmake file and separate it into different targets: dev, test and
prod. See readme to know what the difference is.

Summary:
This commit includes
1. testing through catch in the browser
2. improved cmake
3. O3 optimization for production builds