Use setTimeout instead of setInterval
Advantage: no need to call clearInterval anymore.
This now looks very alike to a while loop, but we ask the browser to
call the function again as soon, as we get into the next event loop.

Another advantage is that setting a timeout and let it call immediately
should be more efficient and predictable. I imagine with setInterval we
could get into situations, where the interpreter's get_answer is slower
than the frequency of the interval. God knows what would happen then ...