Make time_limit_test even harder
Tries to fix same problem as described in dbe886e.
We suspect that the problem is lacking time precision in browsers
and the faster your computer is, the more likely it becomes that
the solution is found within the precision of the timer,
leading the test to fail.
Still failed for me, so making it even harder ...